refactor(workflow): isolate runtime cleanup mechanics
